Fixed so that table rows are removed when data is deleted on dashboard subviews

+++ resources/js/dashboard.js
@@ -7,32 +7,29 @@ $('input#check-all').click(function() {
 });
 
 $('#remove').on('click', function() {
-    var $checkboxes = $('table#data-table');
-    var $allChecked = $checkboxes.find('input[type="checkbox"]:not(#check-all):checked');
+    var allVals = [];
+    var type = $('input[data-type]').data('type');
 
-    var $type = $('input[data-type]').data('type');
-
-    var ids = [];
-
-    $allChecked.map(function() {
-        return ids.push($(this).data('post-id'));
+    $('input[type="checkbox"]:not(#check-all):checked').each(function() {
+        allVals.push($(this).data('post-id'));
     });
 
-    console.log('ids:');
-    console.log(ids.join(','));
-    console.log('-----');
+    if (allVals.length > 0) {
+        var joined_values = allVals.join(',');
 
-    $.ajax({
-        url: `http://localhost:8000/${$type}`,
-        type: 'DELETE',
-        headers: {'X-CSRF-TOKEN': $('meta[name="csrf-token').attr('content') },
-        data: `ids=${ids}`,
-        success: function(res) {
-            console.log(res);
-            // location.reload();
-        },
-        error: function(res) {
-            console.log(res);
-        }
-    });
+         $.ajax({
+            url: `http://localhost:8000/${type}`,
+            type: 'DELETE',
+            headers: {'X-CSRF-TOKEN': $('meta[name="csrf-token').attr('content') },
+            data: `ids=${joined_values}`,
+            success: function(res) {
+                $('input[type="checkbox"]:not(#check-all):checked').each(function() {
+                    $(this).parents('tr').remove();
+                });
+            },
+            error: function(data) {
+                console.log(data.responseText);
+            }
+        });
+    }
 });
